pnotify-style
зробив все в один js файл
